config: refactored UTMConfiguration into multiple categories
