+++ ConfigurationViews/VMConfigDriveDetailViewController.h
@@ -15,7 +15,7 @@
 //
 
 #import <UIKit/UIKit.h>
-#import "UTMConfiguration.h"
+#import "UTMConfiguration+Drives.h"
 #import "VMConfigViewController.h"
 
 NS_ASSUME_NONNULL_BEGIN
